give Polygons its own demo section
